\name{favorita-package}
\alias{favorita-package}
\alias{favorita}
\docType{package}
\title{
\packageTitle{favorita}
}
\description{
\packageDescription{favorita}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{favorita}
\packageIndices{favorita}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{favorita}

Maintainer: \packageMaintainer{favorita}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
